+++ site/source/contextes/comparateur/domaine/ModeleTravailleurIndependant.tsx
@@ -12,7 +12,7 @@ import {
 	eurosParJour,
 	eurosParMois,
 	MontantRécurrent,
-} from '@/domaine/MontantRécurrent'
+} from '@/domaine/MontantRecurrent'
 import { toOuiNon } from '@/domaine/OuiNon'
 import { DottedName } from '@/domaine/publicodes/DottedName'
 import {
