+++ site/source/contextes/comparateur/hooks/useComparateur.ts
@@ -2,7 +2,7 @@ import * as O from 'effect/Option'
 import { useMemo } from 'react'
 
 import { Montant } from '@/domaine/Montant'
-import { MontantRécurrent } from '@/domaine/MontantRécurrent'
+import { MontantRécurrent } from '@/domaine/MontantRecurrent'
 import { Quantité } from '@/domaine/Quantite'
 
 import { NatureActivité, TypeActivité } from '../domaine/activite'
