+++ site/source/contextes/frontalier-suisse/domaine/cotisation.test.ts
@@ -2,7 +2,7 @@ import * as O from 'effect/Option'
 import { describe, expect, it } from 'vitest'
 
 import { euros } from '@/domaine/MontantPonctuel'
-import { eurosParAn, eurosParMois } from '@/domaine/MontantRécurrent'
+import { eurosParAn, eurosParMois } from '@/domaine/MontantRecurrent'
 import { testerEn } from '@/test/testerEn'
 
 import {
