+++ site/source/contextes/frontalier-suisse/domaine/situation-query-string.test.ts
@@ -1,7 +1,7 @@
 import * as O from 'effect/Option'
 import { describe, expect, it } from 'vitest'
 
-import { eurosParAn } from '@/domaine/MontantRécurrent'
+import { eurosParAn } from '@/domaine/MontantRecurrent'
 
 import {
 	initialSituationFrontalierSuisse,
