+++ site/source/contextes/frontalier-suisse/domaine/situation-query-string.ts
@@ -2,7 +2,7 @@ import { pipe } from 'effect'
 import * as O from 'effect/Option'
 
 import { Montant } from '@/domaine/Montant'
-import { eurosParAn } from '@/domaine/MontantRécurrent'
+import { eurosParAn } from '@/domaine/MontantRecurrent'
 
 import {
 	initialSituationFrontalierSuisse,
