+++ site/source/contextes/économie-collaborative/domaine/location-de-meublé/régime-général.ts
@@ -6,7 +6,7 @@ import {
 	estPlusGrandQue,
 	eurosParAn,
 	moins,
-} from '@/domaine/MontantRécurrent'
+} from '@/domaine/MontantRecurrent'
 
 import {
 	applicableSurRecettesCourteDurée,
