+++ site/source/contextes/économie-collaborative/domaine/location-de-meublé/régime-micro-entreprise.ts
@@ -6,7 +6,7 @@ import {
 	estPlusGrandOuÉgalÀ,
 	estPlusGrandQue,
 	eurosParAn,
-} from '@/domaine/MontantRécurrent'
+} from '@/domaine/MontantRecurrent'
 import {
 	AutoEntrepreneurChiffreAffaireDansPublicodes,
 	AutoEntrepreneurContexteDansPublicodes,
