+++ site/source/contextes/économie-collaborative/domaine/location-de-meublé/situation.ts
@@ -1,7 +1,7 @@
 import * as O from 'effect/Option'
 
 import { Montant } from '@/domaine/Montant'
-import { eurosParAn } from '@/domaine/MontantRécurrent'
+import { eurosParAn } from '@/domaine/MontantRecurrent'
 import { Situation } from '@/domaine/Situation'
 
 interface SituationÉconomieCollaborativeBase extends Situation {
