+++ site/source/pages/assistants/choix-du-statut/rémunération.tsx
@@ -7,7 +7,7 @@ import { usePersistingState } from '@/components/utils/persistState'
 import { Body, H3, InfoButton, MontantField, Strong } from '@/design-system'
 import { ValeurPublicodes } from '@/domaine/engine/PublicodesAdapter'
 import { Montant } from '@/domaine/Montant'
-import { eurosParAn } from '@/domaine/MontantRécurrent'
+import { eurosParAn } from '@/domaine/MontantRecurrent'
 import { DottedName } from '@/domaine/publicodes/DottedName'
 import { enregistreLesRéponsesAuxQuestions } from '@/store/actions/actions'
 import { debounce } from '@/utils'
