+++ site/source/pages/simulateurs/location-de-meublé/documentation/ExempleCalcul.tsx
@@ -3,7 +3,7 @@ import { Trans } from 'react-i18next'
 import { ABATTEMENT_REGIME_GENERAL } from '@/contextes/économie-collaborative'
 import { ExemplePratique } from '@/design-system'
 import { montantToString } from '@/domaine/Montant'
-import { eurosParAn } from '@/domaine/MontantRécurrent'
+import { eurosParAn } from '@/domaine/MontantRecurrent'
 
 export const ExempleCalcul = () => {
 	const taux = ABATTEMENT_REGIME_GENERAL
